+++ packages/riverpod_navigator/lib/src/navigator.dart
@@ -14,7 +14,8 @@ class RNavigator extends RNavigatorCore {
     WidgetBuilder? progressIndicatorBuilder,
   })  : navigatorWraperBuilder = navigatorWraperBuilder ?? NavigatorWraper.new,
         splashBuilder = splashBuilder ?? SplashScreen.new,
-        progressIndicatorBuilder = progressIndicatorBuilder ?? CircularProgressIndicator.new,
+        progressIndicatorBuilder =
+            progressIndicatorBuilder ?? CircularProgressIndicator.new,
         routerDelegate = RRouterDelegate(),
         super(ref, routes) {
     routeInformationParser = RouteInformationParserImpl(pathParser);
@@ -39,6 +40,7 @@ class RNavigator extends RNavigatorCore {
     return screen2Page(segment, (segment) => route.buildScreen(segment));
   }
 
-  IconButton? getAppBarLeading() =>
-      getNavigationStack().length > 1 ? IconButton(icon: Icon(Icons.arrow_back), onPressed: onPopRoute) : null;
+  IconButton? getAppBarLeading() => getNavigationStack().length > 1
+      ? IconButton(icon: Icon(Icons.arrow_back), onPressed: onPopRoute)
+      : null;
 }