+++ packages/riverpod_navigator_core/example/main.dart
@@ -7,29 +7,29 @@ import 'package:riverpod_navigator_core/riverpod_navigator_core.dart';
 class HomeSegment extends TypedSegment {
   const HomeSegment();
   // ignore: avoid_unused_constructor_parameters
-  factory HomeSegment.fromUrlPars(UrlPars pars) => HomeSegment();
+  factory HomeSegment.decode(UrlPars pars) => HomeSegment();
 }
 
 @immutable
 class BookSegment extends TypedSegment {
   const BookSegment({required this.id});
-  factory BookSegment.fromUrlPars(UrlPars pars) =>
+  factory BookSegment.decode(UrlPars pars) =>
       BookSegment(id: pars.getInt('id'));
   final int id;
   @override
-  void toUrlPars(UrlPars pars) => pars.setInt('id', id);
+  void encode(UrlPars pars) => pars.setInt('id', id);
 }
 
 class LoginSegment extends TypedSegment {
   const LoginSegment();
   // ignore: avoid_unused_constructor_parameters
-  factory LoginSegment.fromUrlPars(UrlPars pars) => LoginSegment();
+  factory LoginSegment.decode(UrlPars pars) => LoginSegment();
 }
 
 final routes = <RRouteCore>[
-  RRouteCore<HomeSegment>('home', HomeSegment.fromUrlPars),
-  RRouteCore<BookSegment>('book', BookSegment.fromUrlPars),
-  RRouteCore<LoginSegment>('login', LoginSegment.fromUrlPars),
+  RRouteCore<HomeSegment>('home', HomeSegment.decode),
+  RRouteCore<BookSegment>('book', BookSegment.decode),
+  RRouteCore<LoginSegment>('login', LoginSegment.decode),
 ];
 
 final loginProvider = StateProvider<bool>((_) => false);