+++ packages/riverpod_navigator_core/test/model.dart
@@ -3,36 +3,36 @@ import 'package:riverpod_navigator_core/riverpod_navigator_core.dart';
 class HomeSegment extends TypedSegment {
   const HomeSegment();
   // ignore: avoid_unused_constructor_parameters
-  factory HomeSegment.fromUrlPars(UrlPars pars) => HomeSegment();
+  factory HomeSegment.decode(UrlPars pars) => HomeSegment();
 }
 
 class BookSegment extends TypedSegment {
   const BookSegment({required this.id});
-  factory BookSegment.fromUrlPars(UrlPars pars) =>
+  factory BookSegment.decode(UrlPars pars) =>
       BookSegment(id: pars.getInt('id'));
 
   final int id;
 
   @override
-  void toUrlPars(UrlPars pars) => pars.setInt('id', id);
+  void encode(UrlPars pars) => pars.setInt('id', id);
 }
 
 class LoginSegment extends TypedSegment {
   const LoginSegment();
   // ignore: avoid_unused_constructor_parameters
-  factory LoginSegment.fromUrlPars(UrlPars pars) => LoginSegment();
+  factory LoginSegment.decode(UrlPars pars) => LoginSegment();
 }
 
 final routes = <RRouteCore>[
-  RRouteCore<HomeSegment>('home', HomeSegment.fromUrlPars),
-  RRouteCore<BookSegment>('book', BookSegment.fromUrlPars),
-  RRouteCore<LoginSegment>('login', LoginSegment.fromUrlPars),
+  RRouteCore<HomeSegment>('home', HomeSegment.decode),
+  RRouteCore<BookSegment>('book', BookSegment.decode),
+  RRouteCore<LoginSegment>('login', LoginSegment.decode),
 ];
 
 class TestSegment extends TypedSegment {
   const TestSegment({required this.i, this.s, required this.b, this.d});
 
-  factory TestSegment.fromUrlPars(UrlPars pars) => TestSegment(
+  factory TestSegment.decode(UrlPars pars) => TestSegment(
         i: pars.getInt('i'),
         s: pars.getStringNull('s'),
         b: pars.getBool('b'),
@@ -40,7 +40,7 @@ class TestSegment extends TypedSegment {
       );
 
   @override
-  void toUrlPars(UrlPars pars) =>
+  void encode(UrlPars pars) =>
       pars.setInt('i', i).setString('s', s).setBool('b', b).setDouble('d', d);
 
   final int i;