+++ packages/wikib_utils/lib/src/platform/common.dart
@@ -1,0 +1,11 @@
+// copy also in C:\wikibulary\csharp\auth\authLib\lib.cs
+abstract class Platforms {
+  static const unknown = 'UNKNOWN';
+  static const ios = 'IOS';
+  static const android = 'ANDROID';
+  static const fuschia = 'FUSCHIA';
+  static const windows = 'WINDOWS';
+  static const linux = 'LINUX';
+  static const macos = 'MACOS';
+  static const browser = 'BROWSER';
+}