+++ packages/wikib_utils/lib/wikb_utils.dart
@@ -1,0 +1,4 @@
+export 'src/platform/browser.dart' // implementation for web
+    // ignore: uri_does_not_exist
+    if (dart.library.io) 'src/platform/io.dart';
+export 'src/utils.dart';